Remove HTTP context from log messages.
The intention seems to be that this should only be included when we have
all the information about the request and response (eg. response code).
Including part of the information ourselves (just the request info)
makes the output in Stackdriver difficult to read (it includes 'null'
and 'undefined' in the summary line).

Including the trace ID is enough to get the logs grouped together with
Cloud Run's automatic HTTP request and response log entry, so we can
just skip all of this.